ran cargo fmt

+++ backend/server/src/handler/role.rs
@@ -1,5 +1,5 @@
 use crate::models::app::AppState;
-use crate::models::auth::{AuthUser, OrganisationAdmin, RoleAdmin};
+use crate::models::auth::{AuthUser, RoleAdmin};
 use crate::models::error::ChaosError;
 use crate::models::role::{Role, RoleUpdate};
 use axum::extract::{Json, Path, State};
@@ -11,7 +11,7 @@ pub struct RoleHandler;
 impl RoleHandler {
     pub async fn get(
         State(state): State<AppState>,
-        Path(id): Path<i32>,
+        Path(id): Path<i64>,
         _user: AuthUser,
     ) -> Result<impl IntoResponse, ChaosError> {
         let role = Role::get(id, &state.db).await?;
@@ -20,7 +20,7 @@ impl RoleHandler {
 
     pub async fn delete(
         State(state): State<AppState>,
-        Path(id): Path<i32>,
+        Path(id): Path<i64>,
         _admin: RoleAdmin,
     ) -> Result<impl IntoResponse, ChaosError> {
         Role::delete(id, &state.db).await?;
@@ -29,7 +29,7 @@ impl RoleHandler {
 
     pub async fn update(
         State(state): State<AppState>,
-        Path(id): Path<i32>,
+        Path(id): Path<i64>,
         _admin: RoleAdmin,
         Json(data): Json<RoleUpdate>,
     ) -> Result<impl IntoResponse, ChaosError> {
